+++ cricinfo/output_models/scorecard.py
@@ -9,15 +9,16 @@ from cricinfo.source_models.athelete import Athlete
 from cricinfo.source_models.linescores import LinescorePeriod
 from cricinfo.source_models.match import Match
 from cricinfo.source_models.roster import Player, Roster
-from cricinfo.source_models.team import Team
+from cricinfo.source_models.team import TeamWithColorAndLogos
 
 # ANSI escape codes for colors
 RED = "\033[31m"
 RESET = "\033[0m"
 
-SNAKE_CASE_REGEX = re.compile(r'(?<!^)(?=[A-Z])')
+SNAKE_CASE_REGEX = re.compile(r"(?<!^)(?=[A-Z])")
 
-class HeaderlessTableMixin():
+
+class HeaderlessTableMixin:
     def print_headerless_table(self, rows: list[tuple[str, bool]]):
         table = PrettyTable()
         table.header = False
@@ -33,18 +34,19 @@ class PlayerInningsModel(BaseModel, ABC):
         linescore: LinescorePeriod = data.get("linescore")
         if not linescore:
             return data
-        
+
         for name in args:
-            name_split = str(name).split('.')
+            name_split = str(name).split(".")
             stat_name = name_split[1] if len(name_split) > 1 else name_split[0]
-            data[SNAKE_CASE_REGEX.sub('_', stat_name).lower()] = linescore.gs(name)
+            data[SNAKE_CASE_REGEX.sub("_", stat_name).lower()] = linescore.gs(name)
         return data
-    
+
     def colour_row(self, row_items: list[str], colour: str) -> list[str]:
         return [f"{colour}{cell}{RESET}" for cell in row_items]
-    
+
     @abstractmethod
     def add_to_table(self, table: PrettyTable): ...
+
 
 class BattingInnings(PlayerInningsModel):
     player: Athlete
@@ -55,89 +57,136 @@ class BattingInnings(PlayerInningsModel):
     balls_faced: Optional[int] = None
     fours: Optional[int] = None
     sixes: Optional[int] = None
-    not_out: bool = Field(validation_alias=AliasChoices('not_out', 'notouts'))
+    not_out: bool = Field(validation_alias=AliasChoices("not_out", "notouts"))
 
     @computed_field
     @property
     def player_display(self) -> str:
         return f"{self.player.display_name}{' (c)' if self.captain else ''}{' \u271D' if self.keeper else ''}"
 
-    @model_validator(mode='before')
+    @model_validator(mode="before")
     @classmethod
     def create_batting_attributes(cls, data: dict):
-        data = cls.add_linescore_stats_as_properties(data, "batting.dismissal_text", "runs", "ballsFaced", "notouts", "batting.order", "fours", "sixes")
+        data = cls.add_linescore_stats_as_properties(
+            data,
+            "batting.dismissal_text",
+            "runs",
+            "ballsFaced",
+            "notouts",
+            "batting.order",
+            "fours",
+            "sixes",
+        )
         return data
-    
+
     def add_to_table(self, table: PrettyTable):
-        table.add_row(self.colour_row([self.player_display, 
-                                       self.dismissal_text,
-                                       f"{self.runs}{"*" if self.not_out else ''}", 
-                                       self.balls_faced,
-                                       self.fours,
-                                       self.sixes], 
-                                       RED if self.not_out else RESET))
+        table.add_row(
+            self.colour_row(
+                [
+                    self.player_display,
+                    self.dismissal_text,
+                    f"{self.runs}{"*" if self.not_out else ''}",
+                    self.balls_faced,
+                    self.fours,
+                    self.sixes,
+                ],
+                RED if self.not_out else RESET,
+            )
+        )
+
 
 class BowlingInnings(PlayerInningsModel):
     player: Athlete
-    overs: float|int
+    overs: float | int
     maidens: int
-    runs: int = Field(validation_alias=AliasChoices('runs', 'conceded'))
+    runs: int = Field(validation_alias=AliasChoices("runs", "conceded"))
     wickets: int
 
     @computed_field
     @property
-    def overs_display(self) -> float|int:
+    def overs_display(self) -> float | int:
         return int(self.overs) if self.overs % 1 == 0 else self.overs
 
-    @model_validator(mode='before')
+    @model_validator(mode="before")
     @classmethod
     def create_bowling_attributes(cls, data: dict):
-        return cls.add_linescore_stats_as_properties(data, "overs", "maidens", "conceded", "wickets", "bowling.order")
+        return cls.add_linescore_stats_as_properties(
+            data, "overs", "maidens", "conceded", "wickets", "bowling.order"
+        )
 
     def add_to_table(self, table: PrettyTable):
-        table.add_row([self.player.display_name, self.overs_display, self.maidens, self.runs, self.wickets])
+        table.add_row(
+            [
+                self.player.display_name,
+                self.overs_display,
+                self.maidens,
+                self.runs,
+                self.wickets,
+            ]
+        )
+
 
 class Innings(BaseModel, HeaderlessTableMixin):
     number: int
-    team: Team
+    team: TeamWithColorAndLogos
     batting_score: int
     wickets: int
     batting_description: str
-    batters: list[BattingInnings] = []
-    bowlers: list[BowlingInnings] = []
+    batters: list[BattingInnings] = Field(default_factory=list)
+    bowlers: list[BowlingInnings] = Field(default_factory=list)
 
     @computed_field
     @property
     def score_summary(self) -> bool:
-        wickets_text = f" {self.batting_description}" if self.batting_description == "all out" else f"/{self.wickets}"
+        wickets_text = (
+            f" {self.batting_description}"
+            if self.batting_description == "all out"
+            else f"/{self.wickets}"
+        )
         return f"{self.batting_score}{wickets_text}"
 
     def to_table(self):
-        self.print_headerless_table([(f"Innings {self.number}: {self.team.display_name} {self.score_summary}", False)])
+        self.print_headerless_table(
+            [
+                (
+                    f"Innings {self.number}: {self.team.display_name} {self.score_summary}",
+                    False,
+                )
+            ]
+        )
 
-        self._print_player_innings_table(["", "Dismissal", "Runs", "Balls", "4s", "6s"],
-                                         self.batters, ["", "Dismissal"])
-        
-        self._print_player_innings_table(["", "Overs", "Maidens", "Runs", "Wickets"],
-                                         self.bowlers, [""])
+        self._print_player_innings_table(
+            ["", "Dismissal", "Runs", "Balls", "4s", "6s"],
+            self.batters,
+            ["", "Dismissal"],
+        )
 
-    def _print_player_innings_table(self, field_names: list[str], items: list[PlayerInningsModel],
-                                   field_names_to_left_align: list[str]):
+        self._print_player_innings_table(
+            ["", "Overs", "Maidens", "Runs", "Wickets"], self.bowlers
+        )
+
+    def _print_player_innings_table(
+        self,
+        field_names: list[str],
+        items: list[PlayerInningsModel],
+        field_names_to_left_align: list[str] = None,
+    ):
         table = PrettyTable()
         table.field_names = field_names
-        for name in field_names_to_left_align:
+        for name in field_names_to_left_align or []:
             table.align[name] = "l"
 
         for batter in sorted(items, key=lambda b: b.order):
             batter.add_to_table(table)
         print(table)
-    
+
+
 class Scorecard(BaseModel, HeaderlessTableMixin):
     title: Optional[str]
     summary: Optional[str]
     innings: list[Innings]
 
-    @model_validator(mode='before')
+    @model_validator(mode="before")
     @classmethod
     def create(cls, data: dict):
         match: Match = data["match"]
@@ -145,16 +194,20 @@ class Scorecard(BaseModel, HeaderlessTableMixin):
         data["summary"] = match.header.summary
 
         innings = []
-        for i in range(1,5):
+        for i in range(1, 5):
             team_linescore = match.header.get_batting_linescore_for_period(i)
-            innings.append(Innings(number=i, 
-                                   team=team_linescore[0],
-                                   batting_score=team_linescore[1].runs, 
-                                   wickets=team_linescore[1].wickets,
-                                   batting_description=team_linescore[1].description))
+            innings.append(
+                Innings(
+                    number=i,
+                    team=team_linescore[0],
+                    batting_score=team_linescore[1].runs,
+                    wickets=team_linescore[1].wickets,
+                    batting_description=team_linescore[1].description,
+                )
+            )
         for roster in match.rosters:
             cls._enrich_roster(innings, roster)
-        
+
         data["innings"] = innings
         return data
 
@@ -167,15 +220,16 @@ class Scorecard(BaseModel, HeaderlessTableMixin):
     def _enrich_player(cls, innings: list[Innings], player: Player):
         for linescore in player.linescores:
             if bool(linescore.batted) and bool(int(linescore.batted)):
-                bat = BattingInnings(player=player.athlete,
-                                     captain=player.captain,
-                                     keeper=player.keeper,
-                                     linescore=linescore)
-                innings[linescore.period-1].batters.append(bat)
+                bat = BattingInnings(
+                    player=player.athlete,
+                    captain=player.captain,
+                    keeper=player.keeper,
+                    linescore=linescore,
+                )
+                innings[linescore.period - 1].batters.append(bat)
             elif bool(linescore.bowled) and bool(int(linescore.bowled)):
-                bowl = BowlingInnings(player=player.athlete,
-                                      linescore=linescore)
-                innings[linescore.period-1].bowlers.append(bowl)
+                bowl = BowlingInnings(player=player.athlete, linescore=linescore)
+                innings[linescore.period - 1].bowlers.append(bowl)
 
     def to_table(self):
         self.print_headerless_table([(self.title, True), (self.summary, False)])