+++ cricinfo/source_models/__init__.py
@@ -1,7 +1,7 @@
-from cricinfo.source_models.athelete import AthleteLite, Athlete
-from cricinfo.source_models.common import Ref, Link, Position, CCBaseModel
+from cricinfo.source_models.athelete import AthleteWithFirstAndLastName, Athlete
+from cricinfo.source_models.common import RefMixin, Link, Position, CCBaseModel, PagingModel
 from cricinfo.source_models.match_note import MatchNote
 from cricinfo.source_models.official import Official
 from cricinfo.source_models.roster import Roster
-from cricinfo.source_models.team import Team
+from cricinfo.source_models.team import TeamWithColorAndLogos
 from cricinfo.source_models.venue import Venue