+++ cricinfo/source_models/team.py
@@ -1,8 +1,15 @@
-from cricinfo.source_models.common import Ref, CCBaseModel
+from abc import ABC
 
-class Team(CCBaseModel):
+from cricinfo.source_models.common import CCBaseModel, NameMixin, RefMixin
+
+
+class TeamCommon(CCBaseModel, ABC):
     id: str
     abbreviation: str
     display_name: str
+
+class TeamWithName(TeamCommon, NameMixin): ...
+
+class TeamWithColorAndLogos(TeamCommon):
     color: str
-    logos: list[Ref]
+    logos: list[RefMixin]