+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2]
+### Added
+- Output play-by-play commentary data
+
+### Changed
+- Generate requirements files using `pip-compile`
+
 ## [0.0.1]
 ### Added
 - Output a scorecard from the JSON data of a match